A little bit of code cleanup
