+++ strategies_test_helpers_test.go
@@ -1,0 +1,6 @@
+package trader
+
+func mkClose(close float64) Candle {
+	toP := func(x float64) Price { return Price(x*float64(PriceScale) + 0.5) }
+	return Candle{Close: toP(close)}
+}
